James Brown has a sat-on-z customer that pointed out the md5crypt script doesn't work because grub doesn't exist on s390.
Brad Hinson looked over the scripts, and noticed the only point of calling grub in the md5crypt script is to use the built-in md5crypt grub shell command.  So he replaced this with a call to openssl (see attached patch).  With this, we really don't need the md5crypt script any more, and it works for everyone..