Fix localization error messages for "Satellite Administrators" key.
Satellite was string replaced with spacewalk, but looks like "Satellite Administrators" is still being placed in the db during initial org creation. Suspect this value is looked up by the code and then a corresponding string resource key is expected to translate it. Because the string is being treated as a key, opted to leave it as such and just add the corresponding key but resolve it to the correct "Spacewalk Administrators".