Delete unused quickstart.jsp and dead code it exposes.
Looks like this was an RHN specific page despite having Satellite references replaced with Spacewalk. Deleting it, all the sitenav's that reference it (note: all using acl is not satellite), and what I suspect is a dead create user action in struts-config.xml from the hosted days as well. (unable to find any references to it in java or web, and there's a satellite version of the same action clearly visible below).