Makefile.git: do not allow minor releases for package that pack source as .tar.gz.
As we are the upstream of packages that together are Spacewalk, we should
make sure that other projects are not confused when they talk about for
example

	spacewalk-package-0.5.tar.gz

As tag-release and tag-minor-release targets tag with the full
name-version-release, if we do minor releases for packages that pack
source as tar.gz, we will be generating and potentially releasing
multiple tar.gzs with different contents. This would lead to confusion
and unhappiness, hence this added check.