Makefile.git: reset release to 1 upon make tag-release, not to 0.
Devan pointed out

http://fedoraproject.org/wiki/PackageMaintainers/CreatingPackageHowTo#Spec_file_pieces_explained:

Release: The initial value of the release should normally be
"1%{?dist}". Then, increment the number every time you release a new
package for the same version of software. If a new version of the
software being packaged is released, the version number should be
changed to reflect the new software version, and the release number
should be reset to 1.