Revert "Revert "Merge branch 'multiorg' of ssh://git.fedorahosted.org/git/spacewalk""
This reverts commit 3dbe0924eab839069355e7b10cbcedb191d34ced.

Conflicts:

	java/code/src/com/redhat/rhn/domain/org/Org.java
	java/code/src/com/redhat/rhn/domain/org/TrustSet.java

Reason for this commit is as follows, there was an accidental merge from multiorg branch into master when the opposite was intended. This commit was then reverted in master.

When the time came to actually merge master into multiorg, git encounters the "revert" commit just like any other commit and applies it, effectively undoing all the work in multiorg branch and thus, causing many problems. This is "reverting" the "revert" and correcting a couple small conflicts that resulted from subsequent work on a couple java classes.