Removing velocity from install.common.jar.depenedencies.
Fix to help get spacewalk-java building in brew. Velocity jars originate in the tempjars location but were being lumped into an ant variable that resulted in them appearing in a call to build-jar-repository, which couldn't find the correct library to symlink them to. (see build.xml resolve-local call to jpackage-deps) Immediately after this we attempt to actually copy them, although the build probably wasn't getting this far.