Makefile.git: make it possible to call make srpm even if the directory of the package has changed.
We now call git archive with $(GIT_COMMIT_ID):$(GIT_PREFIX),
instead of just $(GIT_COMMIT_ID) and relying on the directory of the
package being still the same as the one in rel-eng/packages/$(NAME).
That however means that git archive get tree's SHA1, not commit's.
In that case git archive puts current time as the modification time
to the tar, not the time of the commit.

That's we we need to postprocess the tar and change the modification
times and the comment in the header with tar-fixup-stamp-comment.pl.