Commit of rhn-satellite-schema-5.1.0-27 schema sources.
Putting those $Id$ values back, to be able to reproduce universe.satellite.sql.

In the Spacewalk 0.1 yum repository, we've published package
rhn-satellite-schema-5.1.0-27.noarch.rpm, taken from Satellite 5.1.0.
The rpm does not correspond to any commit in Spacewalk's git
repository. This is an attempt to recreate the history:

This commit corresponds to Satellite SVN commit 138602 of schema/, with
files that do not appear in the initial commit
	c43c7764d22ca8a78fd6f446b0892b6dec5e78a8
nor in the these rhn-satellite-schema-5.1.0-27 sources removed, plus
code sanitized, using the r172472 change.

The universe.satellite.sql from rhn-satellite-schema-5.1.0-27.noarch.rpm
can be recreated from this commit using

    	cd schema/satellite && \
    	make -f ../build/Makefile SCHEMA=satellite VERSION=5.1.0 \
    		RELEASE=27 schema-clean schema-release