462499 -  RHEL 5.2 yum-rhn-plugin is broken after 4.2.3 update If we get an exception when getting repo xml file, refresh our up2dateAuth and retry
This fixes an issue where a proxy and client get into an odd state
of the client having an up2dateAuth cached, yet the proxy has no data
cached for the client.  A call to up2dateAuth.updateLoginInfo() is required
so the call flows through the proxy and data for the client is now available.