adding errata.findByCve() api
Did not add a junit test.  The reason being we do not currently have a way via the UI or API to associated a CVE with an Errata.  We can create the errata, but not do that association.  Did add a python test case that uses valid data for RHEL; however, the user can change the properities to provide a valid CVE prior to running.