Changing /var/log/rhn ownership perms in proxy spec.
On previous Satellite releases this directory was owned by root:apache with permissions of 770. Four packages claim ownership to this directory after spacewalk is installed, but three originate here and one in backend. Backend has the correct permissions, these appeared to be wrong.

This is now checked for in spacewalk-setup but figured we may as well correct this to the proper values here.