Remove permission setting on /var/log/rhn.
Going to tackle this problem from another angle, suspect the bad permissions are a result of one of the 4 or so packages that own the dir having bad permissions in the spec file. Updated the proxy spec and rebuilt (which is where I suspect the bad perms are coming from), will see if this corrects the issue rather than an explicit call here.