459411: Workaround for aparent Firefox textfield bug.
Firefox appears to be populating staticNetworkIf field with the name of the logged in user on the page load if you have saved credentials for the Spacewalk/Satellite instance.

Workaround to force it back to empty string if the field is disabled using Javascript.

Also includes fix for the field not being properly disabled after submitting the form.