apis: move several kickstart apis to kickstart.profile package
Moved several of the apis (added in spacewalk 0.1-0.3) from the kickstart package to the kickstart.profile package.  The intent is to have all of those profile-related apis in the same package.

This included:
setKickstartTree
setChildChannels
listScripts
addScript
removeScript
downloadKickstart
getPartitioningScheme
setPartitioningScheme

Since the apis to create a profile have been in an official product, I did not move those (or the corresponding delete).  I may deprecate them in a later submit and move them to kickstart.profile also.