456472 - UI updates to ActivationKey->Packages to allow for inclusion of package architecture
This commit includes quite a lot of changes, primarily as the functionality
for ActivationKeys->Packages was ported from perl to java and enhanced to enable
users to specify an optional package architecture (e.g. libjpeg.i386,
a2ps.x86_64, vim-X11).

In addition, modifications were made to the XMLRPC APIs to account for these
enhancements.  This included the following on the activationkey handler:
- getDetails()
  - packages - added to return
  - packageNames - plan to deprecate on the return
- listActivationKeys()
  - packages - added to return
  - packageNames - plan to deprecate on the return
- addPackages() - added api
- removePackages() - added api
- addPackageNames() - plan to deprecate api
- removePackageNames() - plan to deprecate api

Plan to provide another commit where the above mentioded deprecations will
be marked as such w/in the api docs.