fixing issue where system.listSystemEvents was returning everything as strings (even what was supposed to be ints).
In the process, i removed the old Mapify stuff that converted a bean to a struct and changed the logic of the call such that instead of getting a list of Action objects and then looking up the ServerAction object for the server (connected with that Action), it instead just looks up the ServerAction object directly.  Which makes more sense to me.