build.py: Fix issues with packages using build.py.props and --tag.
Can now build such packages from any location in the git checkout, provided the
package had a build.py.props at that point in time. Will tag new versions of
these packages for the new file shortly. If something older is needed we can
either git-checkout a past commit ID and use Makefiles, or add a --props
parameter to force a properties file to be used.