477228 - ActionManager cancelAction deleting server action too early
The ActionManager.cancelAction was deleting the ServerAction before it was finished using it.  Basically, it would delete the ServerActions then perform action.getServerActions resulting in an empty list.  When this empty list was passed to KickstartFactory.failKickstartSessions, the hibernate query could generate an exception because it expected the servers collection (i.e. list passed) to have contents.