479738 - update Help to point to the help page
Updating the site navigations to point the Help to the help page. Currently, it links to www.redhat.com/docs/manuals/satellite.  Also updated links to documents not delivered as part of Spacewalk to point to the www.redhat.com link mentioned.  The purpose for this change is to enable the user to have a way to get to the document search capability and api docs.