480233 - update the ProfileEntry to include package arch id
The hibernate mapping for the rhnServerProfilePackage was missing the mapping for the package_arch_id.  This was causing an ISE to be generated when a profile was deleted using the WebUI.  The reason this occured was that without the arch_id hibernate was grabbing multiple records (if there were multiple pkgs of different arch), when it only expected to receive one.  This caused hibernate to throw an exception.