build.py: Add --untagged-diffs option.
Prints each package with differences between it's most recent tag and
HEAD, as well as the actual diff.

Very similar to git-untagged-commits.pl which was proving insufficient
in branches with a complicated history due to the way it looks only for
SHA1's in one branch but not the other. (cherry-pick and merging divergent
branches can result in SHA1s for code that is actually in both)