Spacewalk::Setup: we do not want to stop the embedded database.
The trouble is that running stop will run listener while it
is not configured, so it will put logs in
/opt/apps/oracle/web/product/10.2.0/db_1/listener.log
and /opt/apps/oracle/web/product/10.2.0/db_1/log, instead to
some subdirectory.

Maybe if the previous installation failed we need to do something
about it but we prefer to to check first. Removing for now.