476331 - install.pl: replace selinux_enabled test with have_semodule and getenforce.
SELinux support for Spacewalk/Satellite is done via SELinux policy
modules.  Therefore, it is not provided on systems that do not support
modular SELinux policy (RHEL 4 and earlier).  On those systems, we
"support" Disabled or Permissive, not Enforcing.  We check for
/usr/sbin/semodule -l to figure out what system we run on.

On systems that support SELinux modules (RHEL 5), we want the
installation to only proceed (by default) if SELinux is not Disabled.
The reason is that if the installation is done with SELinux Disabled,
modules will not be loaded and files and directories won't be
labeled, and there will be extra steps needed after installation
before re-enabling SELinux.