Changed required version number to 1.4.2 for both KOAN and Cobbler per bz486186
Note that in java/spacewalk-java.spec in particular there remains some odd (IMO)
syntax wrt Requires statements:   One sees 'Requires: <pkg> >= 0:<some version>'

I've come across no documentation for this.  mpdehaan said he suspects (but
is not certain) that it's related to "epoch" but is regardless firmly of the
belief that it has no proper place in relation to cobbler or koan.

I'm leaving other occurences of the 0:<blah> syntax in place for lack of
a better understanding.  Hopefully someone will have some context
at some point - it'd be nice to know what's meant by this (or at least
why it is that the spec files are in fact building at all).