kickstart session status where ks is syncing to a profile has an correct url to the stored profile...
If a kickstart profile includes a stored profile that should be synced during the kickstart, when the kickstart is being done on a system the session status page includes a url to allow the user to look at the stored profile.  Unfortunately, that link was currently broken.  It was pointing to a pxt page; however, the stored profile pages were ported to java.