build.py: More dist-cvs support changes.
Add a DummyBuilder for packages that we can't build rpms for. (i.e.
because their sources were too big to add to the SCM and are instead
stored in a lookaside directory)

Make build.py callable in repos that don't track a copy of it's source.
(nor the perl scripts we rely on)