Fix Fedora 10 rpm headers installation error.
Code was checking if monitoring should be supported by looking for
NPalert rpm to be installed on the system. Due to changes either in
Python or in rpm-python, __nonzero__ is no longer defined on the rpm.hdr
class. Changed code to see if a header was found or not instead.