480191 - fix query used by systems->software->packages->install
The query used by this page was returning packages that were not actually
available to the given system.  For example, for an i386 system it was listing
AMD64 packages, for an x86_64 system it was listing i386 packages that were
not part of the x86_64 channels.  In order to address this, added the package
arch to the inner select used to retrieve packages associated with the system.