487234 - fix the query used by the system->software->upgrade page
The query to display the packages that are available for upgrade for a system
was recently updated to support inclusion of package arch; however, the query
had an error in it that caused the result set to be far larger than expected.
For example, on an i386 system that had 116 possible packages to be upgraded
it was returning 1594 packages.  In those, it would repeat the same package
several times (e.g. libjpeg.i386, libjpeg.i386...) as well as list packages
that did not apply for the arch of the system (e.g. libjpeg.x86_64).