488548 - api - org.migrateSystems - fixes related to reactivationkeys, custom info and configuration files
This commit addresses a few fixes that were needed to the migration of systems
between orgs.  This included:
1. Remove the reactivation key associated with the system.  This is necessary
   because the reactivation key is assoicated with the org it was created
   in (i.e. the originating org)
2. Remove any custom info values that were defined for the system.  This is
   necessary because the key associated with the value remains in the
   originating org.
3. Migrate any configuration files that were associated with the system.
   This basically moves any config files and revisions that are either
   Locally-Managed or in the Local Sandbox for the system, to the new org.

Also, updated the migration so that while performing the migration the system's
current entitlement settings are ignored.  This is necessary because it is
possible to have a system with Provisioning where config files are created
and then Provisioning is removed.  In that case, the config files still, but
cannot be accessed; however, they need to be moved with the system.