489968 - updated query used for System->Software->Install
With the existing query (before changes), if an x86_64 system had an i386
package installed, the software list did not show the x86_64 package.
The reason was the query would omit any package from the list, if there
was a package with the same 'name' already installed.  This needed to be
extended to be name+arch.