491021 - api - ensure that repomd regeneration is triggered by APIs
When APIs are invoked that alter the packages contained within a channel,
repomd needs to be triggered.  This is accomplished by invoking the
ChannelManager.queueChannelChange API that was added as part of bz 468781.

While investigating bz 491021, I found that modifications have already been
made that will enable this to be triggered for various APIs; therefore, this
commit is much smaller than initially expected... :)

Below is a quick summary:

mespace packages:

- removePackage - uses PackageManager.schedulePackageRemoval which already
  invokes ChannelManager.queueChannelChange

namespace channel.software:

- addPackages - uses ChannelEditor addPackages->changePackages which already
  invokes ChannelManager.queueChannelChange

- clone - updated NewChannelHelper.clone to queueChannelChange for the newly
  cloned channel

- create - n/a, since no packages associated at channel creation time..

- delete - uses ChannelManager.deleteChannel which already invokes
  ChannelManager.queueChannelChange

- mergePackages - updated ChannelSoftwareHandler.mergePackages to invoke
  queueChannelChange

- removePackages - uses ChannelEditor removePackages->changePackages which
  already invokes ChannelManager.queueChannelChange

namespace errata:
- There aren't any changes needed at this time to the current errata APIs.
  The reason being the APIs for errata do not alter the content
  of a channel (ie. no package adds/removes)

Also, updated ErrataFactory.publishErrataPackagesToChannel to invoke
ChannelManager.queueChannelChange.  This is used when a package is added to
a channel as a result of associating the package with Errata.