489616 - update NewChannelHelper.clone to create channel family, if none exists
In most cases, a private channel family will exist before a channel is
cloned.  For example, recent changes were introduced to ensure that a private
channel family is created for scenarios like initial activation of the server,
syncing of channels as well as creation of organizations; however, there is an
edge case that is not covered.

The edge case involves earlier installations (e.g. 0.1) that are upgraded to
the latest release (e.g. 0.5).  For example, if a user has Spacewalk 0.1
installed and they sync a channel, no private family is created.
If the user then upgrades to the latest release that contains the other
improvements, there still will not exist a private channel family.
As a result, if the user then attempts to clone the channel that they
synced in 0.1 using the API (channel.software.clone) a null pointer
exception will be generated.

This commit is to address this edge case...