493421 - api - kickstart.tree.deleteTreeAndProfiles fixed to delete associated profiles
The API as initially written appeared to be a copy/paste of
kickstart.tree.delete.  As a result, if a tree had any profiles associated
with it, the request was rejected.  This commit includes updates to remove
any profiles associated with the tree.  Also, added a junit test for the API
which was previously missing...