491668 - update Spacewalk Apache conf to support .htaccess
The current httpd.conf file that is deployed as part of Apache has the following
configuration:

<Directory "/var/www/html">
    Options Indexes FollowSymLinks
    AllowOverride None
    Order allow,deny
    Allow from all
</Directory>

The value of "None" for AllowOverride basically disables the support for
the .htaccess files within the application directories that reside under
/var/www/html.  This is a problem for Spacewalk because we need the
.htaccess files to be processed and supported; otherwise, several of
the perl pages are affected.

Unfortunately, the above Directory block is processed after the
zz-spacewalk-www.conf and zz-spacewalk-server.conf files are read in.
As a result, this block overrides the current zz configurations.

In order to avoid making modifications to the standard httpd.conf,
this commit will add a new entry to the zz conf that will basically
'fake' out Apache by creating a separate document with a path
of /var/www/html/*.