handle orphaned GID's the same way as orphaned UID's
Hi,

Below is a patch to rhncfgcli_verify.py to make it handle orphaned
GID's the same way as orphaned UID's are handled. I recently found out
the current implementation raises a KeyError exception if an orphaned
GID is encountered. This happens when rhncfg_client verify is run on a
system on which one of the deployed configuration files is owned by a
group that doesn't exist (anymore).

Patch is basically an adaptation of the code that handles UID's in the
same file, about ten lines up.

I usually spend my time being an sysadmin, so I hope I'm doing this
'sending patch' stuff the right way ;-)

Hope this helps.

Maxim Burgerhout
maxim@wzzrd.com
----------------
GPG Fingerprint
1CC2 A9B2 FE2E 799D 01DB 8A89 0AE8 B60A ACA3 4452