492206-Fix for Kickstart Profile Template parse error.
Cobbler uses cheetah to generate the actual kickstart file from a template. Previously the user was NOT notified about
cheetah parse errors when the customer used the raw kickstart file create/edit. Or in other words, if the user entered
bad constructs like #if with no condition causing cheetah syntax errors, the user would not be warned that there were
parse errors. This fix adds that warning and points him to the view kickstart page which lists the actual error.
This fix also contains fixes on DownloadUtils to work properly with cobbler 1.6.3. In cobbler 1.6 the wget on bad profile
raises a 500 (bad as in parse errors not bad is in profile Not found..) causing url.openStream() to raise an IOEexception
there by preventing proper lookup.. This code has now been fixed to use HttpURL connection to check the response code
first and get the error stream content if 500 was raised.. A new DownloadException will be raised if there is a 500 or
any other response code other than a 200 ... The content of the errorStream is also captured in the exception..