488603 - KickstartFileFormat fix, to deal with blank space interpreter
When generating a post script in a KS wizard if the user enters space as the
interpreter type, the kickformatter assumes that the kickstart had a valid
interpreter and includes the --interpreter tag. This has now been fixed in
this commit. Had to change StringUtils.isEmpty to StringUtils.isBlank in a
couple of places...