config: Remove duplicate symlinks to httpd crt and key.
Appears we were trying to create two sets of symlinks pointing to the
same httpd server.crt and server.key. The links were to appear in
/etc/pki/tls/, one pair named server.* and one pair named spacewalk.*.
The server.* pair was not getting installed, likely due to the %ghost
directive in the spec file. The spacewalk pair was. Can't see any reason
we need both so removing the pair that wasn't even getting installed.