494673 - update system migration to support null config channels
Prior to these changes, the "Server" object would create a local config
channel any time the user invoked Server.getLocalOverride or
Server.getSandboxOverride.  In the case of a server migrating between
orgs, we don't want that behavior.  We simply want to see if the system
has a local config channel and if so, migrate it with the server as it
goes to the new org.  These changes will enable this by adding methods
to get the override channels or 'null' if one does not exist.