443500 - Changed logic to determine packages to remove to include the server's current package information.
In dealing with 494794 (package removals for multiple servers in the SSM should only create one event) I came across this issue. The query to determine which packages should be removed from each server simply used the entire list of selected packages selected by the user (the Java query) or all packages associated with the action (the Python query).

This caused a problem when dealing with multiarch as the client doesn't check the architecture when deleting the package. The change here is to be more precise in determining which packages should be deleted by taking into account the information we have on current packages installed on the server (i.e. rhnServerPackage).

In the process, I cleaned up the Python query which was unnecessarily joining with the rhnPackage and rhnChannelPackage tables. Those changes were not related to the bug fix but rather just general cleanup of the query.