496719 - api - update apis that have map input to generate error for invalid keys
The following APIs take a 'map' as an input argument:
    activationkey.addPackages
    activationkey.removePackages
    activationkey.setDetails
    channel.software.clone
    configchannel.createOrUpdatePath
    errata.create
    errata.setDetails
    system.setDetails
    system.config.createOrUpdatePath
    system.provisioning.snapshot.deleteSnapshots (2 APIs)
    system.provisioning.snapshot.listSnapshots
    user.setDetails

For these APIs, if the user supplies an invalid key value, the key is
currently ignored.  This commit will throw an exception to the user if they
supply an invalid key value.  This is important because a user could
accidentally provide an incorrectly spelled key and not know.

e.g.
key='descripton'
value='this is the description'

In this case, since user mispelled description in the key, the entry
would be ignored.