496080 - Fix channel with package lookup to filter on org.
Previously the code was paying no attention to whether or not the user
had access to the channel we searched for a package in, because it was
usually a tools channel with rhncfg we were looking for, we didn't hit
this issue. If however another org had cloned the tools channel, you
could get a channel not found error adding virt platform entitlements to
a system. The situation is even more likely to surface in Spacewalk.

Modified to check that the channel's org ID is null, equal to the users
org id, or there's a suitable trust + sharing relationship going on.