309601 - add md5crypt/satpasswd/satwho to the spacewalk-backend-tools
This commit is actually several commits squashed together.  The end result
is the move of the existing scripts from spacewalk/scripts/user-accounts
to spacewalk/backend/satellite_tools/accounts and updates to include them
along with manpages for satpasswd and satwho within the spacewalk-backend-tools
pkg.

309601 - adding md5crypt/satpasswd/satwho to the spaceawlk-backend-tools

These scripts are currently in git under scripts/user-accounts; however,
we'd like to bundle them as part of spacewalk-backend-tools so that they
are accessible by root on the spacewalk server as part of installation.

309601 - minor update to Makefile

309601 - Makefile fixes

309601 - Makfile updates

309601 - remove original scripts, add man pages and update makefile to support generating man pages