495847 - New ListTag 3 functionality to add selected servers to SSM.
This was added as a decorator that can be applied to any ListTag table that represents a list of selectable servers. It also includes an option to remove any existing servers from the SSM prior to adding the selected servers.

The hooks for the logic are in the BaseSetHelper; no changes are necessary to the actions.