489902 - fix help links to work with rhn-il8n-guides
This commit should fix the help links to point to the correct locations in the
rhn-il8n-guides.  These help links still need to be refactored to work with
Spacewalk using rhn_docs.conf which will point to www.redhat.com manuals.  This
will be done as part of a future commit.