498275 - api - system.obtainReactivationKey updated to replace existing key
In the UI, if a user generates a reactivation key, the new key replaces the
old one; however, with the current API it simply creates a new key.  As
a result, there will be multiple reactivation keys for the system; however,
only one is seen from the UI.

This update will remove any existing reactivation keys for the system before
creating the new one.