433660 - Removed the restriction in the UI that prevents orgs with 0 entitlements from being shown on the org page of an entitlement.
I added a new call to the manager in parallel with the existing call that filters out 8 entitlement orgs because the existing one is already being used in an XMLRPC API call. Currently, the new call is only used by the UI.

Also made a change to ChannelFamily's getMaxMember method to not use the equals method but rather compare by ID. I was seeing some weirdness with the Orgs not indicating equality when it should have. I'm not fully sure why; the OrgImpl equals method looks sound. Will have to investigate further.