499258 - update Alter Channel Subscriptions to not ISE when base channel is changed
With a recent change to SystemManager.unsubscribeServerFromChannel, when it
is invoked from UpdateBaseChannelCommand.store, a hibernate flush is performed.

This flush is generating an exception indicating the following:
Caused by: org.hibernate.NonUniqueObjectException: a different object with the
same identifier value was already associated with the session:
[com.redhat.rhn.domain.server.Server#1000014288]

This commit will ensure that a hibernate reload is instead executed thereby
generating the expected result (i.e. no ISE and base channel is updated...)