Refactor Schema Makefiles
The initial build takes longer but subsequent builds work as
the should where only those items that have changed will build.

Add sample upgrade for 0.4-0.5.  This is an example that splits
the common items and references (via include) the forked things
that are created as CREATE OR REPLACE.  The postgres stuff aren't
ported by I wanted to add a placeholder.