437361 - Added all orgs (except default org) to the entitlement's org subtab.
Added a new call into OrgManager for returning entitlements with empty orgs. There is currently an XMLRPC API call that uses the existing method. I'll file a new BZ to address that (similar to the issue seen in the software entitlement calls) in an effort to deprecate the existing call.