499515 - fix ISEs with Solaris patch install/remove and cluster install
The modifications for multi-arch changed a query to include arch_id and
added a new query to handle the case where an arch_id was not available;
however, the Solaris code for doing patch install/remove and patch cluster
install were not updated to use the "no arch" query.  This commit addresses
that.

Also, while testing, noticed that when applying a patch cluster, the
"action message" on the page to inform the user that the action was scheduled
was showing up twice on the page.  In order to address that, fixed the
struts config to do a redirect to the next action.