497119 - channel changes - update to use hibernate refresh vs reload
In serveral places within the code, when a base or child channel association
was changed (e.g. public - protected), the logic would invoke
SystemManager.unsubscribeServerFromChannel with a flush=true which
would cause the logic to perform a hibernate flush and evict of the
session.  This would generate errors (ISE).  In order to address this,
updated the logic to invoke the method so that it will instead perform
a hibernate refresh.

Ran tests before and after to confirm the ISE was generated before the
changes, but not after and that the data is updated as expected.