Fix PostgreSQL clearing of the DB during setup.
Originally we were dropping the 'public' schema and recreating it, but
we later began using schemas as "packages" and this method of clearing
the db no longer worked. Replaced to drop the database configured and
re-created it.

This code will probably we useful when we start looking to do "embedded"
PostgreSQL installations.