500688 - Clarify --run-updater help info in spacewalk-setup.
Three states possible, undefined means to prompt the user during
install, yes means to automatically install missing deps, and no means
to terminate if any deps are missing.

Also noticed here that install.pl had a copy of the command line help
data that didn't seem to be used. Removed it and tested, the --help
still displays properly, so taking this out. Note that install.pl gets
it's help data from Setup.pl library, installed as part of
spacewalk-setup as well as on the iso's under install/lib/.