Fix tests and actions broken by recent change to set clearing logic.
Commit 20da1583321f488edf13d24f62ed1d4b872c2cf1 pushed some set clearing
logic up to a base class which caused a number of test failures, but
also it exposed that some actions actually depend on the set NOT being
cleared.

For now fixing the tests and adjusting the actions in question to
override the clear set boolean getter, as I suspect defaulting to
clearing the set is still the safest thing to do. If this becomes
a larger problem this will have to be reverted, as well as
20da1583321f488edf13d24f62ed1d4b872c2cf1.