502099 - api - update systemgroup.addOrRemoveAdmins to not allow changes to access for sat/org admins
Updated the api to throw an exception if the user attempts to change the
admin access permissions associated with a sat/org admin for a given
system group.  These users are granted that permission by default and
it should not be changed.  (Note: this is to make the behavior consistent
with the UI.)