502641 - rhnSystemMigrations remove not null constraints from to/from org ids
Prior to these changes, the rhnSystemMigrations had a not null constraint on
the to_org_id and from_org_id columns.  This would result in an ISE from the UI when attempting to delete an org whose org id was in one of these columns
because those columns also have a constraint to set them to null when the org
is deleted.  As a result, these 2 constraints conflicted.