501358 - api - channel.software.create - update to provide more detail on label/name errors
Currently, when executing the API, if an error exists with either channel label
or name, the exception message is similar to:
	Invalid channel label
	Invalid channel name

Recently, modifications were made to the UI to provide more descriptive
information for these name/label errors.  As a result, updating the API to
use the same/similar messages.