Fixes to support mod_jk >= 2.2.26.
Using a newer version of mod_jk (which we only use on RHEL 4 installs)
resulted in apache not being able to communicate with tomcat. This was
caused by a change in 2.2.26 where mod_jk apache directives must be
defined per vhost, rather than globally. To correct we must specify
JkMountCopy in the vhost definitions in ssl.conf (which we modify during
spacewalk-setup) as well as zz-spacewalk-server.conf.

Because of apache loading order issues, this commit introduces
00-spacewalk-mod_jk.conf to load mod_jk and configure it early.