501224 - api - enhance system.listSystemEvents to include more detail on events executed
Currently, when a user looks at the event history in the UI, they see
more information than they can currently obtain with the API.

For example,
- for a package action (e.g. install/remove/upgrade/verify)
  - user can see the pkgs that were involved in the event
- for an errata update
  - user can see the advisory name & synopsis
- for a config file deploy/upload
  - user can see the file name and possibly revision info
- for a config file compare/diff
  - user can see the files compared and the differences found

The above information is not returned in the API; therefore, enhancing the
system.listSystemEvents to include an optional 'additional_info' array
that will include this information when it is available.

One thing to note is that these details are stored in different tables
depending on the action that was executed.  As a result, we have to
perform different queries depending on the action type; therefore, rather
than use a basic serializer, the api handler itself will gather the results
and populate the data to be returned.