503642 - update KickstartScheduleCommand to store action id after the action has been saved to the db
This change was needed to address a null pointer exception that occured
whenever a user would execute the XMLRPC API system.provisionSystem.

In the past, the Action was saved immediately following scheduleKickstartAction;however, in a commit earlier in the release, this save was moved further down
in the logic.  This is Ok; however, the logic which stored the action id in
the object (i.e. this.kickstartActionId) was not moved.  As a result, the
action id was stored as null, because the action id does not exist until
the Action is saved to the database.