503090 - Exclude rhnlib from kickstart profile sync.
Latest version of rhnlib is hacked into place with wget and rpm -Uvh in
kickstart post script, so we will always have the latest version anyhow.
It appears that this can cause the database to not know rhnlib is
installed, and on RHEL 4 when the package sync is run it attempts to
install rhnlib fresh causing an error. (as it's already installed)