503801 - update channel details edit to not refresh package cache
Performing an edit of the basic channel details (i.e. channels/manage/Edit.do)
such as channel name, description, contact info...etc was resulting in a
refresh of the 'new package' cache.  This isn't needed for the 'edit' since we
are not altering the package content of the channel.  Including this
logic actually causes a timeout at run-time followed by a 500 ISE.