505101 - update css so that links are underlined when hovering
There was a recent change to the css because the documentation used
<a>...</a> but not as links; however, when the mouse would hover over
them, they would be underlined as if they were links.  Unfortunately,
the fix for that caused the behavior of hovering over real links
(e.g. <a href="...">...</a>) to behave inconsistently.  The behavior being
that sometimes hovering would underline the links and other times, it would
not.  (This was mostly observed with the navigation links).  This update
should address that.