498611 - run "semodule -i" in %post and restorecon in %posttrans
According to jnovy there is no guaranteed order of %posttrans :(
Therefore we will need to split %{_sbindir}/%{name}-enable into two parts.
semodule -i should be called in %post when we have guaranteed order due
requires and in %posttrans we can restorecon and there is no need of dependency then.